+++ src/ui/ChannelWidget.py
@@ -16,13 +16,13 @@ class ChannelWidget(QtGui.QWidget, ChannelWidgetUi.Ui_Channel):
     display (and control, where applicable) values, limits and setpoints for
     applied current/light power.
     """
-    def __init__(self, parent, num_channel):
+    def __init__(self, parent, channel_id):
         self.log = logging.getLogger(__name__)
         super(ChannelWidget, self).__init__()
         self.setupUi(self)
         self.parent = parent
-        self.groupBox.setTitle('Channel {0:d}'.format(num_channel))
-        self.num_channel = num_channel
+        self.groupBox.setTitle('Unknown channel')
+        self.num_channel = channel_id
         self.controlled_channel = None
 
         # controls for setpoint
@@ -34,6 +34,8 @@ class ChannelWidget(QtGui.QWidget, ChannelWidgetUi.Ui_Channel):
         self.spin_max.valueChanged[int].connect(self.set_limit)
 
     def initialize(self, channel):
+        assert channel is not None
+        self.groupBox.setTitle(channel.name)
         self.controlled_channel = channel
 
     def refresh(self):
@@ -72,19 +74,17 @@ class ChannelWidget(QtGui.QWidget, ChannelWidgetUi.Ui_Channel):
             self.lbl_set_dbg.setText('{0:0=5.1f}mA'.format(setpoint))
 
     def set_setpoint(self, value):
-        """Changes the value of the setpoint for the controlled channel. Unit depends on
-        CC or CP mode.
+        """Changes the value of the setpoint for the controlled channel.
+        Unit depends on CC or CP mode.
         """
         if self.controlled_channel is not None:
             self.log.debug('updating setpoint of channel {0:d}'.format(self.controlled_channel.id))
             self.controlled_channel.set_setpoint(value/1000.)
 
     def set_limit(self, value):
-        """Changes the value of the limit for the controlled channel. Unit depends on
-        CP or CC mode.
+        """Changes the value of the limit for the controlled channel.
+        Unit depends on CP or CC mode.
         """
         if self.controlled_channel is not None:
             self.log.debug('Setting current limit of channel {0:d}'.format(self.controlled_channel.id))
             self.controlled_channel.set_limit(value/1000.)
-
-
